backend/requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.0
Flask-Migrate>=3.0.0
python-dotenv>=0.19.0 # For loading .env files (optional but good practice)
Werkzeug>=2.0.0 # File uploads, request/response handling
opencv-python-headless # For image processing (no GUI needed)
Pillow # For image manipulation (PIL fork)
pytesseract # Python wrapper for Tesseract OCR engine
tqdm # Progress bars (used in image_enhancer)
requests # Often useful for interacting with other APIs (though not used in core now)
# uuid is built-in, no need to list it here 